requirements.txt
google-adk>=0.3.0
openpyxl>=3.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-cloud-storage>=2.0.0
openai>=1.0.0
pydantic>=2.0.0
